# gc_registry/static/openapi.yaml
components:
  schemas:
...
    GranularCertificateActionResponse:
      properties:
        action_completed_datetime:
          description: The UTC datetime at which the registry confirmed to the User
            that their submitted action had either been successfully completed or
            rejected.
          format: date-time
          title: Action Completed Datetime
          type: string
        action_request_datetime:
          description: The UTC datetime at which the User submitted the action to
            the registry.
          format: date-time
          title: Action Request Datetime
          type: string
        action_response_status:
          description: Specifies whether the requested action has been accepted or
            rejected by the registry.
          title: Action Response Status
          type: string
        certificate_period_end:
          description: The UTC datetime up to which GC Bundles within the specified
            Account are to be filtered.
          format: date-time
          title: Certificate Period End
          type: string
        certificate_period_start:
          description: The UTC datetime from which to filter GC Bundles within the
            specified Account.
          format: date-time
          title: Certificate Period Start
          type: string
        certificate_quantity:
          description: "Overrides GC Bundle range start and end IDs, if specified.\
            \ \n        Of the GC Bundles identified, return the total number of certificates\
            \ to action on, \n        splitting GC Bundles from the start of the range\
            \ where necessary."
          title: Certificate Quantity
          type: integer
        certificate_bundle_status:
          description: Filter on the status of the GC Bundles.
          title: Certificate Status
          type: string
        id:
          description: Filter GC Bundles associated with the specified production
            device.
          format: uuid
          title: Device Id
          type: string
        energy_source:
          description: Filter GC Bundles based on the fuel type used by the production
            Device.
          items:
            type: string
          title: Energy Source
          type: array
        source_id:
          description: The Account ID of the Account within which the action shall
            occur or originate from.
          format: uuid
          title: Source Account Id
          type: string
        source_certificate_certificate_bundle_id_range_end:
          description: If an issuance ID is specified, returns a GC Bundle containing
            all certificates between and inclusive of the range start and end IDs
            provided.
          title: Source Certificate Bundle Id Range End
          type: integer
        source_certificate_certificate_bundle_id_range_start:
          description: If an issuance ID is specified, returns a GC Bundle containing
            all certificates between and inclusive of the range start and end IDs
            provided.
          title: Source Certificate Bundle Id Range Start
          type: integer
        source_certificate_issuance_id:
          description: The specific GC Bundle(s) onto which the action will be performed.
            Returns all GC Bundles with the specified issuance ID.
          format: uuid
          title: Source Certificate Issuance Id
          type: string
        source_id:
          description: The User that is performing the action, and can be verified
            as having the sufficient authority to perform the requested action on
            the Account specified.
          format: uuid
          title: Source User Id
          type: string
        sparse_filter_list:
          additionalProperties:
            format: date-time
            type: string
          description: Overrides all other search criteria. Provide a list of Device
            ID - Datetime pairs to retrieve GC Bundles issued to each Device and datetime
            specified.
          title: Sparse Filter List
          type: object
      required:
      - source_id
      - source_id
      - action_response_status
      title: GranularCertificateActionResponse
      type: object
...
    GranularCertificateCancellation:
      properties:
        action_completed_datetime:
          description: The UTC datetime at which the registry confirmed to the User
            that their submitted action had either been successfully completed or
            rejected.
          format: date-time
          title: Action Completed Datetime
          type: string
        action_id:
          description: A unique ID assigned to this action.
          format: uuid
          title: Action Id
          type: string
        action_request_datetime:
          description: The UTC datetime at which the User submitted the action to
            the registry.
          format: date-time
          title: Action Request Datetime
          type: string
        beneficiary:
          description: The Beneficiary entity that may make a claim on the attributes
            of the cancelled GC Bundles. If not specified, the Account holder is treated
            as the Beneficiary.
          title: Beneficiary
          type: string
        certificate_period_end:
          description: The UTC datetime up to which GC Bundles within the specified
            Account are to be filtered.
          format: date-time
          title: Certificate Period End
          type: string
        certificate_period_start:
          description: The UTC datetime from which to filter GC Bundles within the
            specified Account.
          format: date-time
          title: Certificate Period Start
          type: string
        certificate_quantity:
          description: "Overrides GC Bundle range start and end IDs, if specified.\
            \ \n        Of the GC Bundles identified, return the total number of certificates\
            \ to action on, \n        splitting GC Bundles from the start of the range\
            \ where necessary."
          title: Certificate Quantity
          type: integer
        certificate_bundle_status:
          description: Filter on the status of the GC Bundles.
          title: Certificate Status
          type: string
        id:
          description: Filter GC Bundles associated with the specified production
            device.
          format: uuid
          title: Device Id
          type: string
        energy_source:
          description: Filter GC Bundles based on the fuel type used by the production
            Device.
          items:
            type: string
          title: Energy Source
          type: array
        source_id:
          description: The Account ID of the Account within which the action shall
            occur or originate from.
          format: uuid
          title: Source Account Id
          type: string
        source_certificate_certificate_bundle_id_range_end:
          description: If an issuance ID is specified, returns a GC Bundle containing
            all certificates between and inclusive of the range start and end IDs
            provided.
          title: Source Certificate Bundle Id Range End
          type: integer
        source_certificate_certificate_bundle_id_range_start:
          description: If an issuance ID is specified, returns a GC Bundle containing
            all certificates between and inclusive of the range start and end IDs
            provided.
          title: Source Certificate Bundle Id Range Start
          type: integer
        source_certificate_issuance_id:
          description: The specific GC Bundle(s) onto which the action will be performed.
            Returns all GC Bundles with the specified issuance ID.
          format: uuid
          title: Source Certificate Issuance Id
          type: string
        source_id:
          description: The User that is performing the action, and can be verified
            as having the sufficient authority to perform the requested action on
            the Account specified.
          format: uuid
          title: Source User Id
          type: string
        sparse_filter_list:
          additionalProperties:
            format: date-time
            type: string
          description: Overrides all other search criteria. Provide a list of Device
            ID - Datetime pairs to retrieve GC Bundles issued to each Device and datetime
            specified.
          title: Sparse Filter List
          type: object
      required:
      - source_id
      - source_id
      title: GranularCertificateCancellation
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: token
      type: oauth2
info:
  contact:
    email: connor@futureenergy.associates
    name: Please direct feedback to
  description: "\nThis document outlines the EnergyTag API Specification to accompany\
    \ Version 2 of \nthe EnergyTag Standard.\nThe functionality outlined within each\
    \ API interaction represents the recommended \nminimum required detail necessary\
    \ to implement a consistent Granular Certificate \n(GC) registry system. \nWithin\
    \ the Certificates section, the definition of GCs as specified in the Create \n\
    Certificate POST request, alongside the filtering parameters required to instigate\
    \ \nquery, transfer, and cancellation requests, will form core components of the\
    \ standard \nand impact the functionality of the registry the greatest. Subsequent\
    \ sections \ndescribing Organisation, User, Account, and Device management are\
    \ **recommendations \nonly** and do not form part of the standard, although terminology\
    \ used therein is \nreferenced in the certificate definitions. They are included\
    \ in the database class \ndiagram for the sake of completeness in representing\
    \ the relationships between the \ncertificate and action definitions; the fields\
    \ implemented for these tables are \nleft to the discretion of the registry operator.\
    \ \n<br>\n\nWhere Universally Unique Identifiers (UUIDs) have been used, these\
    \ can be replaced with \nany consistent and unique identification mechanism preferred\
...
openapi: 3.1.0
paths:
...
  /certificates/certificates/cancel:
    post:
      description: Cancel a fixed number of certificates matched to the given filter
        parameters within the specified Account.
      operationId: certificate_bundle_cancellation_certificates_certificates_cancel_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateCancellation'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Cancellation
      tags:
      - Certificates
  /certificates/certificates/claim:
    post:
      description: 'Claim a fixed number of cancelled certificates matching the provided
        search criteria within a given Account,

        if the User is specified as the Beneficiary of those cancelled GCs. For more
        information on the claim process,

        please see page 15 of the EnergyTag GC Scheme Standard document.'
      operationId: certificate_bundle_claim_certificates_certificates_claim_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateClaim'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Claim
      tags:
      - Certificates
  /certificates/certificates/create:
    post:
      description: Create a GC Bundle with the specified properties.
      operationId: create_certificate_bundle_certificates_certificates_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateBundleBase'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateBundle'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Certificate Bundle
      tags:
      - Certificates
  /certificates/certificates/query:
    get:
      description: Return all certificates from the specified Account that match the
        provided search criteria.
      operationId: query_certificate_bundles_certificates_certificates_query_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateQuery'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateQueryResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Query Certificate Bundles
      tags:
      - Certificates
  /certificates/certificates/recurring_cancel:
    post:
      description: Set up a protocol that cancels a fixed number of certificates matching
        the provided search criteria within a given Account once per time period.
      operationId: certificate_bundle_recurring_cancellation_certificates_certificates_recurring_cancel_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateRecurringCancellation'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Recurring Cancellation
      tags:
      - Certificates
  /certificates/certificates/recurring_transfer:
    post:
      description: Set up a protocol that transfers a fixed number of certificates
        matching the provided search criteria to a given target Account once per time
        period.
      operationId: certificate_bundle_recurring_transfer_certificates_certificates_recurring_transfer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateRecurringTransfer'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Recurring Transfer
      tags:
      - Certificates
...
  /certificates/certificates/transfer:
    post:
      description: Transfer a fixed number of certificates matched to the given filter
        parameters to the specified target Account.
      operationId: certificate_bundle_transfer_certificates_certificates_transfer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateTransfer'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Transfer
      tags:
      - Certificates
...
  /certificates/certificates/withdraw:
    post:
      description: (Issuing Body only) - Withdraw a fixed number of certificates from
        the specified Account matching the provided search criteria.
      operationId: certificate_bundle_withdraw_certificates_certificates_withdraw_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GranularCertificateWithdraw'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GranularCertificateActionResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Certificate Bundle Withdraw
      tags:
      - Certificates
...